Move RoutingId to routing package
